[RHSM] Add missing server url
For the moment, if we want to use RHSM with username and password
AND custom serverl url, for example 'stage', then we fail.
In this case we go to the prod server with creds from the stage,
which is bug.
So, add server url to the appropriate task to make it work
with any server different from the default one (prod).